Fix printf parameters in OpenCollada.

+++ src/LightsprintIO/OpenCollada/COLLADAFramework/src/COLLADAFWValidate.cpp
@@ -56,8 +56,8 @@ namespace COLLADAFW
 			if (verbose)
 				printf("ERROR: [%s] Found %d input values for %d keys\n",
 				animationCurve->getName().c_str(),
-				animationCurve->getInputValues().getValuesCount(),
-				keyCount);
+				(int)animationCurve->getInputValues().getValuesCount(),
+				(int)keyCount);
 			failure_count +=1;
 		}
 
@@ -77,8 +77,8 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] Found %d interpolation types for %d keys\n",
 					animationCurve->getName().c_str(),
-					animationCurve->getInterpolationTypes().getCount(),
-					keyCount);
+					(int)animationCurve->getInterpolationTypes().getCount(),
+					(int)keyCount);
 				failure_count +=1;
 			}
 
@@ -104,7 +104,7 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] Found %d mixed interpolation types (expected only one type).\n",
 					animationCurve->getName().c_str(),
-					animationCurve->getInterpolationTypes().getCount());
+					(int)animationCurve->getInterpolationTypes().getCount());
 				failure_count +=1;
 			}
 		}
@@ -120,8 +120,8 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] Found %d IN tangent values for %d tangents\n",
 					animationCurve->getName().c_str(),
-					animationCurve->getInTangentValues().getValuesCount(),
-					tangentCount);
+					(int)animationCurve->getInTangentValues().getValuesCount(),
+					(int)tangentCount);
 			failure_count +=1;
 		}
 
@@ -130,8 +130,8 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] Found %d OUT tangent values for %d tangents\n",
 					animationCurve->getName().c_str(),
-					animationCurve->getOutTangentValues().getValuesCount(),
-					tangentCount);
+					(int)animationCurve->getOutTangentValues().getValuesCount(),
+					(int)tangentCount);
 			failure_count +=1;
 		}
 
@@ -154,8 +154,8 @@ namespace COLLADAFW
 			if (verbose)
 				printf("ERROR: [%s] found %d bind matrices and %d joints\n",
 				skinControllerData->getName().c_str(),
-				skinControllerData->getInverseBindMatrices().getCount(),
-				jointsCount);
+				(int)skinControllerData->getInverseBindMatrices().getCount(),
+				(int)jointsCount);
 			failure_count +=1;
 		}
 	
@@ -174,8 +174,8 @@ namespace COLLADAFW
 			if (verbose)
 				printf("ERROR: [%s] found %d joint-Vertex Pairs and %d weights\n",
 				skinControllerData->getName().c_str(),
-				jointsVertexPairCount,
-				weightIndices.getCount());
+				(int)jointsVertexPairCount,
+				(int)weightIndices.getCount());
 			failure_count +=1;
 		}
 
@@ -185,9 +185,9 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] weight index %d=%d points outside of weight array of length %d\n",
 					skinControllerData->getName().c_str(),
-					i,
-					weightIndices[i],
-					weightsCount);
+					(int)i,
+					(int)weightIndices[i],
+					(int)weightsCount);
 				failure_count +=1;
 			}
 		}
@@ -199,8 +199,8 @@ namespace COLLADAFW
 			if (verbose)
 				printf("ERROR: [%s] found %d joint-Vertex Pairs and %d joint indices\n",
 				skinControllerData->getName().c_str(),
-				jointsVertexPairCount,
-				jointIndices.getCount());
+				(int)jointsVertexPairCount,
+				(int)jointIndices.getCount());
 			failure_count +=1;
 		}
 
@@ -210,9 +210,9 @@ namespace COLLADAFW
 				if (verbose)
 					printf("ERROR: [%s] joint index %d=%d points outside of weight array of length %d\n",
 					skinControllerData->getName().c_str(),
-					i,
+					(int)i,
 					(unsigned int)abs(jointIndices[i]),
-					jointsCount);
+					(int)jointsCount);
 				failure_count +=1;
 			}
 		}
